Remove unused CheckResponse.Error
This field described only authz-specific errors and was filled in
envoy.Handler.Check() response when also returning non-nil error, but in
such case the grpc framework was not using the response anyway.
This field was also used for metrics, but no codepath was actually
setting them, as flowcontrol never set these.

Also:
* Create errors using grpc/status package, so that we can control the
  grpc status.
* Add missing sampled logs for error conditions.

Drive-by:
* Remove unused error from ClassifierEngine.Classify(), as it's
  infallible (all errors are reported individually per-label).
* Remove unused code from authz.go.