Added support for changing agent installation mode without re-creation of CR (#1089)
### Description of change

User has to uninstall and recreate the Agent CR currently to change the
installation mode to Sidecar.
With this change, they can just modify the existing CR to change the
installation mode and operator will handle the rest.
Restarting all the pods on changing the installation is not feasible so
added event in the CR to indicate the same.

Fixes #896