blueprint: revert integrator from latency gradient policy (#1205)
### Description of change

There is a possible issue with the way we are using the integrator in
the latency gradient policy. Specifically, we need to figure out whether
the min value has to be applied or not. I am reverting to earlier
circuit for now and we can re-add integrator next week.

##### Checklist

- [X] Tested in playground or other setup
- [ ] Documentation is changed or added
- [ ] Tests and/or benchmarks are included
- [ ] Breaking changes

<!-- Reviewable:start -->
- - -
This change is [<img src="https://reviewable.io/review_button.svg"
height="34" align="absmiddle"
alt="Reviewable"/>](https://reviewable.io/reviews/fluxninja/aperture/1205)
<!-- Reviewable:end -->