Fix handling of EDITOR and VISUAL environment variables with spaces (#1376)
Only EDITOR variable was checked, and it was assumed that the value is just a command to be executed,
but it can be both command and its arguments (e.g. `EDITOR="emacsclient -c"`. This commit adds support
for that, as well as support for VISUAL environment variable.

<!-- Reviewable:start -->
- - -
This change is [<img src="https://reviewable.io/review_button.svg"
height="34" align="absmiddle"
alt="Reviewable"/>](https://reviewable.io/reviews/fluxninja/aperture/1376)
<!-- Reviewable:end -->