Add clusterIP to entity
In this change,
1. we get the service that the endpoint belongs to
2. extract clusterIP if it is not "" or "None" (for headless services)
3. create entry in entitycache with this IP

Refactor k8s service-discovery

In this change, we move away from the using watcher interface towards informer interface. While watching using watcher interface, there is a chance to miss events. In the informer interface, it handles regularly listing and maintaining an internal cache. Because of this, we can do away with our intermediary cache for holding watched entities.