Enforce single controller per project via leader election (#2644)
Now the controller won't start until it wins a leader election (note:
it's a separate leader election than the one agents are doing). The
`Election` was refactored so that it has configurable key and works for
both controllers' and agents' elections.

Drive-by:
* Simplified election package by removing trackers dependency.
* Added some structured concurrency helpers in `panichandler` (and used
when appropriate):
* `WaitGroup` – like `conc.WaitGroup`, but uses our panichandler, No
more `wg.Add(1)` and `defer wg.Done()`.
* `CancellableWaitGroup` – a `WaitGroup` with context and cancellation.
Also provides `GoOnDone` to schedule a callback when a given channel is
done (closed).

Resolves #2588