Upload images to dockerhub for dev changes (#2784)
**New Feature:**
- Added a new parameter `dockerhub-ops-image` to the CircleCI workflows.
This allows for specifying a Docker image name for pushing to the
`fluxninjaops` DockerHub account. If this parameter is provided, the
workflow will tag and push the image to the `fluxninjaops` DockerHub
account. This feature has been implemented for the `aperture-agent`,
`aperture-controller`, and `aperture-operator` workflows.

**Style:**
- Updated the log message in `controller.go` from "Failed to close
controller connection" to "Fail to close controller connection". This
change does not affect the functionality of the code.