Made version mandatory for aperturectl install and uninstall (#2802)
We cannot publish a chart with version as latest as it doesn't consider
it as a valid semver.

- **Change**: The `--version` flag in the `aperturectl install` and
`aperturectl uninstall` commands now requires explicit input from the
user. The default value of "latest" has been removed.
- **Impact**: This change enhances the precision of the installation and
uninstallation process, ensuring users have full control over the
version of Aperture they are managing. It prevents unintentional
installations or uninstallations of the latest version when a specific
version is intended.
- **Documentation**: Updated command references in the documentation to
reflect this change. Users are now guided to provide a specific version
when using these commands.