Make the policy hash stable (#2790)
Policy hashing logic was changed to not depend on aperture version. The policy spec is
now hashed directly – as stored as JSON in etcd. This ensures that after
applying policy, its hash will never change.

The sha256 was chosen so that it's easy to compute manually, in case of
debugging.

Note: Because of change in hashing logic, all policies will now report new
hashes, so it's a breaking change wrt. policy hashes. Hopefully, a last one.